Pull, show and delete models

Pull, show and delete models

Usage

pull_model(model = NULL, server = NULL, insecure = FALSE)

show_model(model = NULL, server = NULL)

delete_model(model, server = NULL)

copy_model(model, destination = paste0(model, "-copy"), server = NULL)

Arguments

model

name of the model. Defaults to "llama2" when NULL (except in delete_model).

server

URL to an Ollama server (not the API). Defaults to "http://localhost:11434".

insecure

allow insecure connections to the library. Only use this if you are pulling from your own library during development. description

destination

name of the copied model.

Value

(invisible) a tibble with information about the model (except in delete_model)

Details

pull_model(): downloads model

show_model(): displays information about a local model

copy_model(): creates a model with another name from an existing model

delete_model(): deletes local model

Model names: Model names follow a model:tag format, where model can have an optional namespace such as example/model. Some examples are orca-mini:3b-q4_1 and llama2:70b. The tag is optional and, if not provided, will default to latest. The tag is used to identify a specific version.

Examples

if (FALSE) {
model_info <- pull_model("mixtral")
# after you pull, you can get the same information with:
model_info <- show_model("mixtral")
}