Chat with a LLM through Ollama

Chat with a LLM through Ollama

Usage

query(
  q,
  model = NULL,
  screen = TRUE,
  server = NULL,
  images = NULL,
  model_params = NULL,
  format = NULL,
  output = c("httr2_response", "text", "list", "data.frame", "httr2_request"),
  template = NULL
)

chat(
  q,
  model = NULL,
  screen = TRUE,
  server = NULL,
  images = NULL,
  model_params = NULL,
  template = NULL
)

Arguments

q

the question as a character string or a conversation object.

model

which model(s) to use. See https://ollama.com/library for options. Default is "llama3.1". Set option(rollama_model = "modelname") to change default for the current session. See pull_model for more details.

screen

Logical. Should the answer be printed to the screen.

server

URL to an Ollama server (not the API). Defaults to "http://localhost:11434".

images

path(s) to images (for multimodal models such as llava).

model_params

a named list of additional model parameters listed in the documentation for the Modelfile such as temperature. Use a seed and set the temperature to zero to get reproducible results (see examples).

format

the format to return a response in. Currently the only accepted value is "json". This is a parameter send to Ollama to influence how the model behaves. To select an output format use the output parameter.

output

what the function should return. Possible values are "httr2_response", "text", "list", "data.frame" or "httr2_request" (in case you want to perform requests yourself).

template

the prompt template to use (overrides what is defined in the Modelfile).

Value

an httr2 response.

Details

query sends a single question to the API, without knowledge about previous questions (only the config message is relevant). chat treats new messages as part of the same conversation until new_chat is called.

Examples